13 - make preprocessing skip files that are too short

+++ 13_rvc_surgery/train_v4.py
@@ -115,10 +115,10 @@ class RVCTrainingModule(pl.LightningModule):
                 batch['pitch'] = MyFeatures.f0_to_coarse(batch['pitch_fine']).squeeze(0) # Recalculate coarse
             with torch.no_grad():
                 o, x_mask, z_stats = self.net_g.infer(
-                    batch['rvc_feat'].to(self.device), 
+                    batch['rvc_feat'].to(self.device).to(self.dtype), 
                     batch['lengths'].to(self.device), 
                     batch['pitch'].to(self.device), 
-                    batch['pitch_fine'].to(self.device),
+                    batch['pitch_fine'].to(self.device).to(self.dtype),
                     batch['sids'].to(self.device),
                     noise_scale=self.config.train.noise_scale_test
                 )
